move message processor to separate mod
